Fix space security
