+++ packages/presentation/src/components/SpacesPopup.svelte
@@ -13,21 +13,24 @@
 // limitations under the License.
 -->
 <script lang="ts">
-  import type { Class, Doc, DocumentQuery, FindOptions, Ref, Space } from '@hcengineering/core'
+  import { Class, Doc, DocumentQuery, FindOptions, Ref, Space, getCurrentAccount } from '@hcengineering/core'
   import { AnySvelteComponent, ButtonSize } from '@hcengineering/ui'
   import { ObjectCreate } from '../types'
-  import ObjectPopup from './ObjectPopup.svelte'
+  import { createQuery } from '../utils'
+  import DocPopup from './DocPopup.svelte'
   import SpaceInfo from './SpaceInfo.svelte'
 
   export let _class: Ref<Class<Space>>
   export let selected: Ref<Space> | undefined
-  export let spaceQuery: DocumentQuery<Space> | undefined
+  export let spaceQuery: DocumentQuery<Space> | undefined = {}
   export let spaceOptions: FindOptions<Space> | undefined = {}
   export let create: ObjectCreate | undefined = undefined
   export let size: ButtonSize = 'large'
   export let allowDeselect = false
   export let component: AnySvelteComponent | undefined = undefined
   export let componentProps: any | undefined = undefined
+
+  let search: string | undefined = undefined
 
   $: _create =
     create !== undefined
@@ -36,19 +39,40 @@
           update: (doc: Doc) => (doc as Space).name
         }
       : undefined
+
+  const me = getCurrentAccount()._id
+
+  const query = createQuery()
+  $: query.query(
+    _class,
+    {
+      ...(spaceQuery ?? {}),
+      ...(search !== undefined && search !== ''
+        ? {
+            name: { $like: `%${search}%` }
+          }
+        : {})
+    },
+    (res) => {
+      spaces = res.filter((p) => !p.private || p.members.includes(me))
+    },
+    spaceOptions
+  )
+
+  let spaces: Space[] = []
 </script>
 
-<ObjectPopup
+<DocPopup
   {_class}
-  options={spaceOptions}
+  objects={spaces}
   {selected}
-  bind:docQuery={spaceQuery}
   multiSelect={false}
   {allowDeselect}
   shadows={true}
   create={_create}
   on:update
   on:close
+  on:search={(e) => (search = e.detail)}
 >
   <svelte:fragment slot="item" let:item={space}>
     {#if component}
@@ -57,4 +81,4 @@
       <SpaceInfo {size} value={space} />
     {/if}
   </svelte:fragment>
-</ObjectPopup>
+</DocPopup>
