+++ packages/presentation/src/index.ts
@@ -26,6 +26,7 @@ export { default as Card } from './components/Card.svelte'
 export { default as MessageBox } from './components/MessageBox.svelte'
 export { default as MessageViewer } from './components/MessageViewer.svelte'
 export { default as ObjectPopup } from './components/ObjectPopup.svelte'
+export { default as DocPopup } from './components/DocPopup.svelte'
 export { default as PDFViewer } from './components/PDFViewer.svelte'
 export { default as SpaceCreateCard } from './components/SpaceCreateCard.svelte'
 export { default as SpaceMultiBoxList } from './components/SpaceMultiBoxList.svelte'
